ArmPlatformPkg/PL011UartLib: Prevent data loss in conversion.
Add an explicit (UINT32) cast to the assignment.

+++ ArmPlatformPkg/Library/PL011UartLib/PL011UartLib.c
@@ -193,7 +193,7 @@ PL011UartInitializePort (
       return RETURN_INVALID_PARAMETER;
     }
 
-    Divisor    = (UartClkInHz * 4) / *BaudRate;
+    Divisor    = (UINT32)((UartClkInHz * 4) / *BaudRate);
     Integer    = Divisor >> FRACTION_PART_SIZE_IN_BITS;
     Fractional = Divisor & FRACTION_PART_MASK;
   }
